requirements.txt
alembic==1.12.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.0.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lesscpy==0.15.1
Mako==1.3.0
MarkupSafe==2.1.3
ply==3.11
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
rcssmin==1.1.2
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
six==1.16.0
SQLAlchemy==2.0.21
stripe==7.8.2
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.3.7
WTForms==3.0.1
xxhash==3.4.1
